profiler/requirements.txt
# python 3.10.14
outlines==0.0.46
langchain-community
unstructured
dspy-ai==2.4.13
weave==0.51.1
optimum==1.21.4 
transformers==4.43.4
auto-gptq==0.7.1
nltk==3.9.1
llama-index==0.11.3
ollama==0.3.2
llama-index-llms-ollama==0.3.0
llama-index-embeddings-ollama==0.3.0
protobuf==3.20
keybert==0.8.5
owlready2==0.47
llama-index-retrievers-bm25==0.4.0
# werkzeug==2.3.8
